Add selection for receipt items

+++ app/src/main/java/com/kjh/posreceiptprinter/MainViewModel.kt
@@ -7,7 +7,7 @@ class MainViewModel : ViewModel() {
     val receipt: MutableLiveData<List<ReceiptItem>> by lazy {
         MutableLiveData(emptyList())
     }
-    var receiptItemId: Int = 0
+    var receiptItemId: Long = 0
 
     val products: Array<String> = Array(12) { "상품 ${it + 1}" }
